@@ -2,7 +2,7 @@ name: webview_flutter_example
 description: Demonstrates how to use the webview_flutter plugin.
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -13,10 +13,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_driver:
-    sdk: flutter
-  integration_test: ^1.0.2+2
-  pedantic: ^1.8.0
 
 flutter:
   uses-material-design: true
