@@ -103,8 +103,8 @@ class WebResourceError {
   /// A user should not need to instantiate this class, but will receive one in
   /// [WebResourceErrorCallback].
   WebResourceError({
-    @required this.errorCode,
-    @required this.description,
+    required this.errorCode,
+    required this.description,
     this.domain,
     this.errorType,
     this.failingUrl,
@@ -131,21 +131,21 @@ class WebResourceError {
   /// in Objective-C. See
   /// https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/ErrorHandlingCocoa/ErrorObjectsDomains/ErrorObjectsDomains.html
   /// for more information on error handling on iOS.
-  final String domain;
+  final String? domain;
 
   /// Description of the error that can be used to communicate the problem to the user.
-  final String description;
+  final String? description;
 
   /// The type this error can be categorized as.
   ///
   /// This will never be `null` on Android, but can be `null` on iOS.
-  final WebResourceErrorType errorType;
+  final WebResourceErrorType? errorType;
 
   /// Gets the URL for which the resource request was made.
   ///
   /// This value is not provided on iOS. Alternatively, you can keep track of
   /// the last values provided to [WebViewPlatformController.loadUrl].
-  final String failingUrl;
+  final String? failingUrl;
 }
 
 /// Interface for talking to the webview's platform implementation.
@@ -176,7 +176,7 @@ abstract class WebViewPlatformController {
   /// Throws an ArgumentError if `url` is not a valid URL string.
   Future<void> loadUrl(
     String url,
-    Map<String, String> headers,
+    Map<String, String>? headers,
   ) {
     throw UnimplementedError(
         "WebView loadUrl is not implemented on the current platform");
@@ -186,7 +186,7 @@ abstract class WebViewPlatformController {
   ///
   /// Any non null field in `settings` will be set as the new setting value.
   /// All null fields in `settings` are ignored.
-  Future<void> updateSettings(WebSettings setting) {
+  Future<void>? updateSettings(WebSettings setting) {
     throw UnimplementedError(
         "WebView updateSettings is not implemented on the current platform");
   }
@@ -194,19 +194,19 @@ abstract class WebViewPlatformController {
   /// Accessor to the current URL that the WebView is displaying.
   ///
   /// If no URL was ever loaded, returns `null`.
-  Future<String> currentUrl() {
+  Future<String?> currentUrl() {
     throw UnimplementedError(
         "WebView currentUrl is not implemented on the current platform");
   }
 
   /// Checks whether there's a back history item.
-  Future<bool> canGoBack() {
+  Future<bool?> canGoBack() {
     throw UnimplementedError(
         "WebView canGoBack is not implemented on the current platform");
   }
 
   /// Checks whether there's a forward history item.
-  Future<bool> canGoForward() {
+  Future<bool?> canGoForward() {
     throw UnimplementedError(
         "WebView canGoForward is not implemented on the current platform");
   }
@@ -250,7 +250,7 @@ abstract class WebViewPlatformController {
   ///
   /// The Future completes with an error if a JavaScript error occurred, or if the type of the
   /// evaluated expression is not supported(e.g on iOS not all non primitive type can be evaluated).
-  Future<String> evaluateJavascript(String javascriptString) {
+  Future<String?> evaluateJavascript(String javascriptString) {
     throw UnimplementedError(
         "WebView evaluateJavascript is not implemented on the current platform");
   }
@@ -281,7 +281,7 @@ abstract class WebViewPlatformController {
   }
 
   /// Returns the title of the currently loaded page.
-  Future<String> getTitle() {
+  Future<String?> getTitle() {
     throw UnimplementedError(
         "WebView getTitle is not implemented on the current platform");
   }
@@ -305,7 +305,7 @@ abstract class WebViewPlatformController {
   /// Return the horizontal scroll position of this view.
   ///
   /// Scroll position is measured from left.
-  Future<int> getScrollX() {
+  Future<int?> getScrollX() {
     throw UnimplementedError(
         "WebView getScrollX is not implemented on the current platform");
   }
@@ -313,7 +313,7 @@ abstract class WebViewPlatformController {
   /// Return the vertical scroll position of this view.
   ///
   /// Scroll position is measured from top.
-  Future<int> getScrollY() {
+  Future<int?> getScrollY() {
     throw UnimplementedError(
         "WebView getScrollY is not implemented on the current platform");
   }
@@ -333,30 +333,30 @@ class WebSetting<T> {
   /// Constructs a setting of the given `value`.
   ///
   /// The [isPresent] field for the instance will be true.
-  WebSetting.of(T value)
+  WebSetting.of(T? value)
       : _value = value,
         isPresent = true;
 
-  final T _value;
+  final T? _value;
 
   /// The setting's value.
   ///
   /// Throws if [WebSetting.isPresent] is false.
-  T get value {
-    if (!isPresent) {
+  T? get value {
+    if (!isPresent!) {
       throw StateError('Cannot access a value of an absent WebSetting');
     }
-    assert(isPresent);
+    assert(isPresent!);
     return _value;
   }
 
   /// True when this web setting instance contains a value.
   ///
   /// When false the [WebSetting.value] getter throws.
-  final bool isPresent;
+  final bool? isPresent;
 
   @override
-  bool operator ==(Object other) {
+  bool operator ==(dynamic other) {
     if (other.runtimeType != runtimeType) return false;
     final WebSetting<T> typedOther = other;
     return typedOther.isPresent == isPresent && typedOther._value == _value;
@@ -386,15 +386,15 @@ class WebSettings {
   }) : assert(userAgent != null);
 
   /// The JavaScript execution mode to be used by the webview.
-  final JavascriptMode javascriptMode;
+  final JavascriptMode? javascriptMode;
 
   /// Whether the [WebView] has a [NavigationDelegate] set.
-  final bool hasNavigationDelegate;
+  final bool? hasNavigationDelegate;
 
   /// Whether to enable the platform's webview content debugging tools.
   ///
   /// See also: [WebView.debuggingEnabled].
-  final bool debuggingEnabled;
+  final bool? debuggingEnabled;
 
   /// The value used for the HTTP `User-Agent:` request header.
   ///
@@ -404,12 +404,12 @@ class WebSettings {
   /// last time it was set.
   ///
   /// See also [WebView.userAgent].
-  final WebSetting<String> userAgent;
+  final WebSetting<String>? userAgent;
 
   /// Whether to allow swipe based navigation in iOS.
   ///
   /// See also: [WebView.gestureNavigationEnabled]
-  final bool gestureNavigationEnabled;
+  final bool? gestureNavigationEnabled;
 
   @override
   String toString() {
@@ -437,12 +437,12 @@ class CreationParams {
   /// The initialUrl to load in the webview.
   ///
   /// When null the webview will be created without loading any page.
-  final String initialUrl;
+  final String? initialUrl;
 
   /// The initial [WebSettings] for the new webview.
   ///
   /// This can later be updated with [WebViewPlatformController.updateSettings].
-  final WebSettings webSettings;
+  final WebSettings? webSettings;
 
   /// The initial set of JavaScript channels that are configured for this webview.
   ///
@@ -455,12 +455,12 @@ class CreationParams {
   /// ```
   // TODO(amirh): describe what should happen when postMessage is called once that code is migrated
   // to PlatformWebView.
-  final Set<String> javascriptChannelNames;
+  final Set<String>? javascriptChannelNames;
 
   /// The value used for the HTTP User-Agent: request header.
   ///
   /// When null the platform's webview default is used for the User-Agent header.
-  final String userAgent;
+  final String? userAgent;
 
   /// Which restrictions apply on automatic media playback.
   final AutoMediaPlaybackPolicy autoMediaPlaybackPolicy;
@@ -505,14 +505,14 @@ abstract class WebViewPlatform {
   ///
   /// `webViewPlatformHandler` must not be null.
   Widget build({
-    BuildContext context,
+    BuildContext? context,
     // TODO(amirh): convert this to be the actual parameters.
     // I'm starting without it as the PR is starting to become pretty big.
     // I'll followup with the conversion PR.
-    CreationParams creationParams,
-    @required WebViewPlatformCallbacksHandler webViewPlatformCallbacksHandler,
-    WebViewPlatformCreatedCallback onWebViewPlatformCreated,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
+    CreationParams? creationParams,
+    @required WebViewPlatformCallbacksHandler? webViewPlatformCallbacksHandler,
+    WebViewPlatformCreatedCallback? onWebViewPlatformCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
   });
 
   /// Clears all cookies for all [WebView] instances.
