@@ -47,10 +47,10 @@ class NavigationRequest {
   NavigationRequest._({this.url, this.isForMainFrame});
 
   /// The URL that will be loaded if the navigation is executed.
-  final String url;
+  final String? url;
 
   /// Whether the navigation request is to be loaded as the main frame.
-  final bool isForMainFrame;
+  final bool? isForMainFrame;
 
   @override
   String toString() {
@@ -79,18 +79,18 @@ enum NavigationDecision {
 class SurfaceAndroidWebView extends AndroidWebView {
   @override
   Widget build({
-    BuildContext context,
-    CreationParams creationParams,
-    WebViewPlatformCreatedCallback onWebViewPlatformCreated,
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
-    @required WebViewPlatformCallbacksHandler webViewPlatformCallbacksHandler,
+    BuildContext? context,
+    CreationParams? creationParams,
+    WebViewPlatformCreatedCallback? onWebViewPlatformCreated,
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers,
+    @required WebViewPlatformCallbacksHandler? webViewPlatformCallbacksHandler,
   }) {
     assert(webViewPlatformCallbacksHandler != null);
     return PlatformViewLink(
       viewType: 'plugins.flutter.io/webview',
       surfaceFactory: (
-        BuildContext context,
-        PlatformViewController controller,
+        BuildContext? context,
+        dynamic controller,
       ) {
         return AndroidViewSurface(
           controller: controller,
@@ -108,7 +108,7 @@ class SurfaceAndroidWebView extends AndroidWebView {
           // directionality.
           layoutDirection: TextDirection.rtl,
           creationParams: MethodChannelWebViewPlatform.creationParamsToMap(
-            creationParams,
+            creationParams!,
           ),
           creationParamsCodec: const StandardMessageCodec(),
         )
@@ -118,7 +118,7 @@ class SurfaceAndroidWebView extends AndroidWebView {
               return;
             }
             onWebViewPlatformCreated(
-              MethodChannelWebViewPlatform(id, webViewPlatformCallbacksHandler),
+              MethodChannelWebViewPlatform(id, webViewPlatformCallbacksHandler!),
             );
           })
           ..create();
@@ -172,8 +172,8 @@ class JavascriptChannel {
   ///
   /// The parameters `name` and `onMessageReceived` must not be null.
   JavascriptChannel({
-    @required this.name,
-    @required this.onMessageReceived,
+    required this.name,
+    required this.onMessageReceived,
   })  : assert(name != null),
         assert(onMessageReceived != null),
         assert(_validChannelNames.hasMatch(name));
@@ -192,7 +192,7 @@ class JavascriptChannel {
   final String name;
 
   /// A callback that's invoked when a message is received through the channel.
-  final JavascriptMessageHandler onMessageReceived;
+  final JavascriptMessageHandler? onMessageReceived;
 }
 
 /// A web view widget for showing html content.
@@ -208,7 +208,7 @@ class WebView extends StatefulWidget {
   ///
   /// The `javascriptMode` and `autoMediaPlaybackPolicy` parameters must not be null.
   const WebView({
-    Key key,
+    Key? key,
     this.onWebViewCreated,
     this.initialUrl,
     this.javascriptMode = JavascriptMode.disabled,
@@ -227,7 +227,7 @@ class WebView extends StatefulWidget {
         assert(initialMediaPlaybackPolicy != null),
         super(key: key);
 
-  static WebViewPlatform _platform;
+  static WebViewPlatform? _platform;
 
   /// Sets a custom [WebViewPlatform].
   ///
@@ -257,11 +257,11 @@ class WebView extends StatefulWidget {
               "Trying to use the default webview implementation for $defaultTargetPlatform but there isn't a default one");
       }
     }
-    return _platform;
+    return _platform!;
   }
 
   /// If not null invoked once the web view is created.
-  final WebViewCreatedCallback onWebViewCreated;
+  final WebViewCreatedCallback? onWebViewCreated;
 
   /// Which gestures should be consumed by the web view.
   ///
@@ -272,13 +272,13 @@ class WebView extends StatefulWidget {
   ///
   /// When this set is empty or null, the web view will only handle pointer events for gestures that
   /// were not claimed by any other gesture recognizer.
-  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers;
 
   /// The initial URL to load.
-  final String initialUrl;
+  final String? initialUrl;
 
   /// Whether Javascript execution is enabled.
-  final JavascriptMode javascriptMode;
+  final JavascriptMode? javascriptMode;
 
   /// The set of [JavascriptChannel]s available to JavaScript code running in the web view.
   ///
@@ -307,7 +307,7 @@ class WebView extends StatefulWidget {
   /// channels in the list.
   ///
   /// A null value is equivalent to an empty set.
-  final Set<JavascriptChannel> javascriptChannels;
+  final Set<JavascriptChannel>? javascriptChannels;
 
   /// A delegate function that decides how to handle navigation actions.
   ///
@@ -331,10 +331,10 @@ class WebView extends StatefulWidget {
   ///     * When a navigationDelegate is set pages with frames are not properly handled by the
   ///       webview, and frames will be opened in the main frame.
   ///     * When a navigationDelegate is set HTTP requests do not include the HTTP referer header.
-  final NavigationDelegate navigationDelegate;
+  final NavigationDelegate? navigationDelegate;
 
   /// Invoked when a page starts loading.
-  final PageStartedCallback onPageStarted;
+  final PageStartedCallback? onPageStarted;
 
   /// Invoked when a page has finished loading.
   ///
@@ -346,13 +346,13 @@ class WebView extends StatefulWidget {
   /// When invoked on iOS or Android, any Javascript code that is embedded
   /// directly in the HTML has been loaded and code injected with
   /// [WebViewController.evaluateJavascript] can assume this.
-  final PageFinishedCallback onPageFinished;
+  final PageFinishedCallback? onPageFinished;
 
   /// Invoked when a web resource has failed to load.
   ///
   /// This can be called for any resource (iframe, image, etc.), not just for
   /// the main page.
-  final WebResourceErrorCallback onWebResourceError;
+  final WebResourceErrorCallback? onWebResourceError;
 
   /// Controls whether WebView debugging is enabled.
   ///
@@ -365,14 +365,14 @@ class WebView extends StatefulWidget {
   /// - From the Menu-bar (of Safari) select Develop -> iPhone Simulator -> <your webview page>
   ///
   /// By default `debuggingEnabled` is false.
-  final bool debuggingEnabled;
+  final bool? debuggingEnabled;
 
   /// A Boolean value indicating whether horizontal swipe gestures will trigger back-forward list navigations.
   ///
   /// This only works on iOS.
   ///
   /// By default `gestureNavigationEnabled` is false.
-  final bool gestureNavigationEnabled;
+  final bool? gestureNavigationEnabled;
 
   /// The value used for the HTTP User-Agent: request header.
   ///
@@ -386,7 +386,7 @@ class WebView extends StatefulWidget {
   /// user agent.
   ///
   /// By default `userAgent` is null.
-  final String userAgent;
+  final String? userAgent;
 
   /// Which restrictions apply on automatic media playback.
   ///
@@ -394,7 +394,7 @@ class WebView extends StatefulWidget {
   /// changes to this parameter are ignored (as long as the state of the [WebView] is preserved).
   ///
   /// The default policy is [AutoMediaPlaybackPolicy.require_user_action_for_all_media_types].
-  final AutoMediaPlaybackPolicy initialMediaPlaybackPolicy;
+  final AutoMediaPlaybackPolicy? initialMediaPlaybackPolicy;
 
   @override
   State<StatefulWidget> createState() => _WebViewState();
@@ -404,7 +404,7 @@ class _WebViewState extends State<WebView> {
   final Completer<WebViewController> _controller =
       Completer<WebViewController>();
 
-  _PlatformCallbacksHandler _platformCallbacksHandler;
+  _PlatformCallbacksHandler? _platformCallbacksHandler;
 
   @override
   Widget build(BuildContext context) {
@@ -429,27 +429,27 @@ class _WebViewState extends State<WebView> {
     super.didUpdateWidget(oldWidget);
     _assertJavascriptChannelNamesAreUnique();
     _controller.future.then((WebViewController controller) {
-      _platformCallbacksHandler._widget = widget;
+      _platformCallbacksHandler!._widget = widget;
       controller._updateWidget(widget);
     });
   }
 
   void _onWebViewPlatformCreated(WebViewPlatformController webViewPlatform) {
     final WebViewController controller =
-        WebViewController._(widget, webViewPlatform, _platformCallbacksHandler);
+        WebViewController._(widget, webViewPlatform, _platformCallbacksHandler!);
     _controller.complete(controller);
     if (widget.onWebViewCreated != null) {
-      widget.onWebViewCreated(controller);
+      widget.onWebViewCreated!(controller);
     }
   }
 
   void _assertJavascriptChannelNamesAreUnique() {
     if (widget.javascriptChannels == null ||
-        widget.javascriptChannels.isEmpty) {
+        widget.javascriptChannels!.isEmpty) {
       return;
     }
-    assert(_extractChannelNames(widget.javascriptChannels).length ==
-        widget.javascriptChannels.length);
+    assert(_extractChannelNames(widget.javascriptChannels!).length ==
+        widget.javascriptChannels!.length);
   }
 }
 
@@ -457,9 +457,9 @@ CreationParams _creationParamsfromWidget(WebView widget) {
   return CreationParams(
     initialUrl: widget.initialUrl,
     webSettings: _webSettingsFromWidget(widget),
-    javascriptChannelNames: _extractChannelNames(widget.javascriptChannels),
+    javascriptChannelNames: _extractChannelNames(widget.javascriptChannels!),
     userAgent: widget.userAgent,
-    autoMediaPlaybackPolicy: widget.initialMediaPlaybackPolicy,
+    autoMediaPlaybackPolicy: widget.initialMediaPlaybackPolicy!,
   );
 }
 
@@ -479,27 +479,27 @@ WebSettings _clearUnchangedWebSettings(
   assert(currentValue.javascriptMode != null);
   assert(currentValue.hasNavigationDelegate != null);
   assert(currentValue.debuggingEnabled != null);
-  assert(currentValue.userAgent.isPresent);
+  assert(currentValue.userAgent!.isPresent!);
   assert(newValue.javascriptMode != null);
   assert(newValue.hasNavigationDelegate != null);
   assert(newValue.debuggingEnabled != null);
-  assert(newValue.userAgent.isPresent);
+  assert(newValue.userAgent!.isPresent!);
 
-  JavascriptMode javascriptMode;
-  bool hasNavigationDelegate;
-  bool debuggingEnabled;
+  JavascriptMode? javascriptMode;
+  bool? hasNavigationDelegate;
+  bool? debuggingEnabled;
   WebSetting<String> userAgent = WebSetting<String>.absent();
   if (currentValue.javascriptMode != newValue.javascriptMode) {
-    javascriptMode = newValue.javascriptMode;
+    javascriptMode = newValue.javascriptMode!;
   }
   if (currentValue.hasNavigationDelegate != newValue.hasNavigationDelegate) {
-    hasNavigationDelegate = newValue.hasNavigationDelegate;
+    hasNavigationDelegate = newValue.hasNavigationDelegate!;
   }
   if (currentValue.debuggingEnabled != newValue.debuggingEnabled) {
-    debuggingEnabled = newValue.debuggingEnabled;
+    debuggingEnabled = newValue.debuggingEnabled!;
   }
   if (currentValue.userAgent != newValue.userAgent) {
-    userAgent = newValue.userAgent;
+    userAgent = newValue.userAgent!;
   }
 
   return WebSettings(
@@ -519,7 +519,7 @@ Set<String> _extractChannelNames(Set<JavascriptChannel> channels) {
 
 class _PlatformCallbacksHandler implements WebViewPlatformCallbacksHandler {
   _PlatformCallbacksHandler(this._widget) {
-    _updateJavascriptChannelsFromSet(_widget.javascriptChannels);
+    _updateJavascriptChannelsFromSet(_widget.javascriptChannels!);
   }
 
   WebView _widget;
@@ -530,15 +530,15 @@ class _PlatformCallbacksHandler implements WebViewPlatformCallbacksHandler {
 
   @override
   void onJavaScriptChannelMessage(String channel, String message) {
-    _javascriptChannels[channel].onMessageReceived(JavascriptMessage(message));
+    _javascriptChannels[channel]!.onMessageReceived!(JavascriptMessage(message));
   }
 
   @override
-  FutureOr<bool> onNavigationRequest({String url, bool isForMainFrame}) async {
+  FutureOr<bool> onNavigationRequest({String? url, bool? isForMainFrame}) async {
     final NavigationRequest request =
         NavigationRequest._(url: url, isForMainFrame: isForMainFrame);
     final bool allowNavigation = _widget.navigationDelegate == null ||
-        await _widget.navigationDelegate(request) ==
+        await _widget.navigationDelegate!(request) ==
             NavigationDecision.navigate;
     return allowNavigation;
   }
@@ -546,21 +546,21 @@ class _PlatformCallbacksHandler implements WebViewPlatformCallbacksHandler {
   @override
   void onPageStarted(String url) {
     if (_widget.onPageStarted != null) {
-      _widget.onPageStarted(url);
+      _widget.onPageStarted!(url);
     }
   }
 
   @override
   void onPageFinished(String url) {
     if (_widget.onPageFinished != null) {
-      _widget.onPageFinished(url);
+      _widget.onPageFinished!(url);
     }
   }
 
   @override
   void onWebResourceError(WebResourceError error) {
     if (_widget.onWebResourceError != null) {
-      _widget.onWebResourceError(error);
+      _widget.onWebResourceError!(error);
     }
   }
 
@@ -585,16 +585,16 @@ class WebViewController {
     this._webViewPlatformController,
     this._platformCallbacksHandler,
   ) : assert(_webViewPlatformController != null) {
-    _settings = _webSettingsFromWidget(_widget);
+    _settings = _webSettingsFromWidget(_widget!);
   }
 
-  final WebViewPlatformController _webViewPlatformController;
+  final WebViewPlatformController? _webViewPlatformController;
 
-  final _PlatformCallbacksHandler _platformCallbacksHandler;
+  final _PlatformCallbacksHandler? _platformCallbacksHandler;
 
-  WebSettings _settings;
+  WebSettings? _settings;
 
-  WebView _widget;
+  WebView? _widget;
 
   /// Loads the specified URL.
   ///
@@ -606,11 +606,11 @@ class WebViewController {
   /// Throws an ArgumentError if `url` is not a valid URL string.
   Future<void> loadUrl(
     String url, {
-    Map<String, String> headers,
+    Map<String, String>? headers,
   }) async {
     assert(url != null);
     _validateUrlString(url);
-    return _webViewPlatformController.loadUrl(url, headers);
+    return _webViewPlatformController!.loadUrl(url, headers);
   }
 
   /// Accessor to the current URL that the WebView is displaying.
@@ -620,43 +620,43 @@ class WebViewController {
   /// current URL changes again by the time this function returns (in other
   /// words, by the time this future completes, the WebView may be displaying a
   /// different URL).
-  Future<String> currentUrl() {
-    return _webViewPlatformController.currentUrl();
+  Future<String?> currentUrl() {
+    return _webViewPlatformController!.currentUrl();
   }
 
   /// Checks whether there's a back history item.
   ///
   /// Note that this operation is asynchronous, and it is possible that the "canGoBack" state has
   /// changed by the time the future completed.
-  Future<bool> canGoBack() {
-    return _webViewPlatformController.canGoBack();
+  Future<bool?> canGoBack() {
+    return _webViewPlatformController!.canGoBack();
   }
 
   /// Checks whether there's a forward history item.
   ///
   /// Note that this operation is asynchronous, and it is possible that the "canGoForward" state has
   /// changed by the time the future completed.
-  Future<bool> canGoForward() {
-    return _webViewPlatformController.canGoForward();
+  Future<bool?> canGoForward() {
+    return _webViewPlatformController!.canGoForward();
   }
 
   /// Goes back in the history of this WebView.
   ///
   /// If there is no back history item this is a no-op.
   Future<void> goBack() {
-    return _webViewPlatformController.goBack();
+    return _webViewPlatformController!.goBack();
   }
 
   /// Goes forward in the history of this WebView.
   ///
   /// If there is no forward history item this is a no-op.
   Future<void> goForward() {
-    return _webViewPlatformController.goForward();
+    return _webViewPlatformController!.goForward();
   }
 
   /// Reloads the current URL.
   Future<void> reload() {
-    return _webViewPlatformController.reload();
+    return _webViewPlatformController!.reload();
   }
 
   /// Clears all caches used by the [WebView].
@@ -670,40 +670,40 @@ class WebViewController {
   ///
   /// Note: Calling this method also triggers a reload.
   Future<void> clearCache() async {
-    await _webViewPlatformController.clearCache();
+    await _webViewPlatformController!.clearCache();
     return reload();
   }
 
   Future<void> _updateWidget(WebView widget) async {
     _widget = widget;
     await _updateSettings(_webSettingsFromWidget(widget));
-    await _updateJavascriptChannels(widget.javascriptChannels);
+    await _updateJavascriptChannels(widget.javascriptChannels!);
   }
 
-  Future<void> _updateSettings(WebSettings newSettings) {
+  Future<void>? _updateSettings(WebSettings newSettings) {
     final WebSettings update =
-        _clearUnchangedWebSettings(_settings, newSettings);
+        _clearUnchangedWebSettings(_settings!, newSettings);
     _settings = newSettings;
-    return _webViewPlatformController.updateSettings(update);
+    return _webViewPlatformController!.updateSettings(update);
   }
 
   Future<void> _updateJavascriptChannels(
       Set<JavascriptChannel> newChannels) async {
     final Set<String> currentChannels =
-        _platformCallbacksHandler._javascriptChannels.keys.toSet();
+        _platformCallbacksHandler!._javascriptChannels.keys.toSet();
     final Set<String> newChannelNames = _extractChannelNames(newChannels);
     final Set<String> channelsToAdd =
         newChannelNames.difference(currentChannels);
     final Set<String> channelsToRemove =
         currentChannels.difference(newChannelNames);
     if (channelsToRemove.isNotEmpty) {
-      await _webViewPlatformController
+      await _webViewPlatformController!
           .removeJavascriptChannels(channelsToRemove);
     }
     if (channelsToAdd.isNotEmpty) {
-      await _webViewPlatformController.addJavascriptChannels(channelsToAdd);
+      await _webViewPlatformController!.addJavascriptChannels(channelsToAdd);
     }
-    _platformCallbacksHandler._updateJavascriptChannelsFromSet(newChannels);
+    _platformCallbacksHandler!._updateJavascriptChannelsFromSet(newChannels);
   }
 
   /// Evaluates a JavaScript expression in the context of the current page.
@@ -722,8 +722,8 @@ class WebViewController {
   /// When evaluating Javascript in a [WebView], it is best practice to wait for
   /// the [WebView.onPageFinished] callback. This guarantees all the Javascript
   /// embedded in the main frame HTML has been loaded.
-  Future<String> evaluateJavascript(String javascriptString) {
-    if (_settings.javascriptMode == JavascriptMode.disabled) {
+  Future<String?> evaluateJavascript(String javascriptString) {
+    if (_settings!.javascriptMode == JavascriptMode.disabled) {
       return Future<String>.error(FlutterError(
           'JavaScript mode must be enabled/unrestricted when calling evaluateJavascript.'));
     }
@@ -734,40 +734,40 @@ class WebViewController {
     // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
     // https://github.com/flutter/flutter/issues/26431
     // ignore: strong_mode_implicit_dynamic_method
-    return _webViewPlatformController.evaluateJavascript(javascriptString);
+    return _webViewPlatformController!.evaluateJavascript(javascriptString);
   }
 
   /// Returns the title of the currently loaded page.
-  Future<String> getTitle() {
-    return _webViewPlatformController.getTitle();
+  Future<String?> getTitle() {
+    return _webViewPlatformController!.getTitle();
   }
 
   /// Sets the WebView's content scroll position.
   ///
   /// The parameters `x` and `y` specify the scroll position in WebView pixels.
   Future<void> scrollTo(int x, int y) {
-    return _webViewPlatformController.scrollTo(x, y);
+    return _webViewPlatformController!.scrollTo(x, y);
   }
 
   /// Move the scrolled position of this view.
   ///
   /// The parameters `x` and `y` specify the amount of WebView pixels to scroll by horizontally and vertically respectively.
   Future<void> scrollBy(int x, int y) {
-    return _webViewPlatformController.scrollBy(x, y);
+    return _webViewPlatformController!.scrollBy(x, y);
   }
 
   /// Return the horizontal scroll position, in WebView pixels, of this view.
   ///
   /// Scroll position is measured from left.
-  Future<int> getScrollX() {
-    return _webViewPlatformController.getScrollX();
+  Future<int?> getScrollX() {
+    return _webViewPlatformController!.getScrollX();
   }
 
   /// Return the vertical scroll position, in WebView pixels, of this view.
   ///
   /// Scroll position is measured from top.
-  Future<int> getScrollY() {
-    return _webViewPlatformController.getScrollY();
+  Future<int?> getScrollY() {
+    return _webViewPlatformController!.getScrollY();
   }
 }
 
@@ -780,7 +780,7 @@ class CookieManager {
 
   CookieManager._();
 
-  static CookieManager _instance;
+  static CookieManager? _instance;
 
   /// Clears all cookies for all [WebView] instances.
   ///
