@@ -89,8 +89,8 @@ class _WebViewExampleState extends State<WebViewExample> {
             onPageFinished: (String url) async {
               var _wvController = await _controller.future;
               //remove any comment
-              double height =
-          double.parse(await _wvController.evaluateJavascript('document.documentElement.scrollHeight;'));
+              var evalJs = await _wvController.evaluateJavascript('document.documentElement.scrollHeight;');
+              double height = double.parse(evalJs!);
               setState(() {
                 _height = height;
               });
@@ -118,7 +118,7 @@ class _WebViewExampleState extends State<WebViewExample> {
               _toasterJavascriptChannel(context),
             ].toSet(),
             navigationDelegate: (NavigationRequest request) {
-              if (request.url.startsWith('https://www.youtube.com/')) {
+              if (request.url!.startsWith('https://www.youtube.com/')) {
                 print('blocking navigation to $request}');
                 return NavigationDecision.prevent;
               }
@@ -153,7 +153,7 @@ class _WebViewExampleState extends State<WebViewExample> {
           if (controller.hasData) {
             return FloatingActionButton(
               onPressed: () async {
-                final String url = await controller.data.currentUrl();
+                final String? url = await controller.data!.currentUrl();
                 // ignore: deprecated_member_use
                 Scaffold.of(context).showSnackBar(
                   SnackBar(content: Text('Favorited $url')),
@@ -192,25 +192,25 @@ class SampleMenu extends StatelessWidget {
           onSelected: (MenuOptions value) {
             switch (value) {
               case MenuOptions.showUserAgent:
-                _onShowUserAgent(controller.data, context);
+                _onShowUserAgent(controller.data!, context);
                 break;
               case MenuOptions.listCookies:
-                _onListCookies(controller.data, context);
+                _onListCookies(controller.data!, context);
                 break;
               case MenuOptions.clearCookies:
                 _onClearCookies(context);
                 break;
               case MenuOptions.addToCache:
-                _onAddToCache(controller.data, context);
+                _onAddToCache(controller.data!, context);
                 break;
               case MenuOptions.listCache:
-                _onListCache(controller.data, context);
+                _onListCache(controller.data!, context);
                 break;
               case MenuOptions.clearCache:
-                _onClearCache(controller.data, context);
+                _onClearCache(controller.data!, context);
                 break;
               case MenuOptions.navigationDelegate:
-                _onNavigationDelegateExample(controller.data, context);
+                _onNavigationDelegateExample(controller.data!, context);
                 break;
             }
           },
@@ -257,7 +257,7 @@ class SampleMenu extends StatelessWidget {
   }
 
   void _onListCookies(WebViewController controller, BuildContext context) async {
-    final String cookies = await controller.evaluateJavascript('document.cookie');
+    final String? cookies = await controller.evaluateJavascript('document.cookie');
     // ignore: deprecated_member_use
     Scaffold.of(context).showSnackBar(SnackBar(
       content: Column(
@@ -265,7 +265,7 @@ class SampleMenu extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
           const Text('Cookies:'),
-          _getCookieList(cookies),
+          _getCookieList(cookies!),
         ],
       ),
     ));
@@ -336,7 +336,7 @@ class NavigationControls extends StatelessWidget {
       future: _webViewControllerFuture,
       builder: (BuildContext context, AsyncSnapshot<WebViewController> snapshot) {
         final bool webViewReady = snapshot.connectionState == ConnectionState.done;
-        final WebViewController controller = snapshot.data;
+        final WebViewController? controller = snapshot.data;
         return Row(
           children: <Widget>[
             IconButton(
@@ -344,7 +344,7 @@ class NavigationControls extends StatelessWidget {
               onPressed: !webViewReady
                   ? null
                   : () async {
-                      if (await controller.canGoBack()) {
+                      if ((await controller!.canGoBack())!) {
                         await controller.goBack();
                       } else {
                         // ignore: deprecated_member_use
@@ -360,7 +360,7 @@ class NavigationControls extends StatelessWidget {
               onPressed: !webViewReady
                   ? null
                   : () async {
-                      if (await controller.canGoForward()) {
+                      if ((await controller!.canGoForward())!) {
                         await controller.goForward();
                       } else {
                         // ignore: deprecated_member_use
@@ -376,7 +376,7 @@ class NavigationControls extends StatelessWidget {
               onPressed: !webViewReady
                   ? null
                   : () {
-                      controller.reload();
+                      controller!.reload();
                     },
             ),
           ],
